Remove Tilo files as symlinks; Fix Entity/ThingDictionary ID bug

+++ tests/LabyrinthOfLore/Entity/ThingDictionaryTest.cpp
@@ -29,7 +29,8 @@ TEST(LabyrinthOfLore_Entity_ThingDictionaryTest, create_new_definition__returns_
    ASSERT_EQ("a rock", retrieved_thing.infer_decorated_name());
 }
 
-TEST(LabyrinthOfLore_Entity_ThingDictionaryTest, create_new_definition__adds_a_new_thing_to_the_dictionary_with_an_id_that_is_at_least_CREATED_THINGS_START)
+TEST(LabyrinthOfLore_Entity_ThingDictionaryTest,
+   create_new_definition__adds_a_new_thing_to_the_dictionary_with_an_id_that_is_at_least_CREATED_THINGS_START)
 {
    LabyrinthOfLore::Entity::ThingDictionary thing_dictionary;
    int new_id_1 = thing_dictionary.create_new_definition({ THING_TYPE_TORCH, "a", "rock" });
@@ -38,8 +39,8 @@ TEST(LabyrinthOfLore_Entity_ThingDictionaryTest, create_new_definition__adds_a_n
 
    ASSERT_EQ(3, thing_dictionary.size());
 
-   EXPECT_EQ(CREATED_THINGS_START+1, new_id_1);
-   EXPECT_EQ(CREATED_THINGS_START+2, new_id_2);
-   EXPECT_EQ(CREATED_THINGS_START+3, new_id_3);
+   EXPECT_EQ((int)CREATED_THINGS_START+1, new_id_1);
+   EXPECT_EQ((int)CREATED_THINGS_START+2, new_id_2);
+   EXPECT_EQ((int)CREATED_THINGS_START+3, new_id_3);
 }
